sexta-feira (0.1)/requirements.txt
SpeechRecognition==3.8.1
vosk==0.3.7
PyAudio--0.2.11
pyttsx3==2.90